add check before rmdir in run stories

+++ tools/src/commands/RunStories.ts
@@ -5,13 +5,14 @@ import inquirer from 'inquirer';
 import path from 'path';
 
 import { runExpoCliAsync } from '../ExpoCLI';
+import Logger from '../Logger';
 
 type Action = {
   platform: 'android' | 'ios' | 'web';
   packageName: string;
 };
 
-async function action({ platform, packageName }: Action) {
+async function action(packageName: string, { platform }: Action) {
   const cwdPkg = require(path.resolve(process.cwd(), 'package.json'));
   const cwdPkgName = cwdPkg.name;
 
@@ -50,6 +51,7 @@ async function action({ platform, packageName }: Action) {
   const targetName = projectName.split('-').join('');
 
   const projectRoot = path.resolve(examplesRoot, projectName);
+
   if (fs.existsSync(projectRoot)) {
     const { shouldRebuild } = await inquirer.prompt({
       type: 'confirm',
@@ -59,21 +61,23 @@ async function action({ platform, packageName }: Action) {
     });
 
     if (!shouldRebuild) {
-      console.log();
-      console.log(`Project found at ${projectRoot}`);
-      console.log();
+      Logger.log();
+      Logger.info(`Project found at ${projectRoot}`);
+      Logger.log();
     }
 
     buildStoryLoader = shouldRebuild;
   }
 
   if (buildStoryLoader) {
-    // @ts-ignore
-    fs.rmdirSync(projectRoot, { recursive: true, force: true });
+    if (fs.existsSync(projectRoot)) {
+      // @ts-ignore
+      fs.rmdirSync(projectRoot, { recursive: true, force: true });
+    }
 
-    console.log();
-    console.log(`🛠  Building fresh story loader for ${packageName}`);
-    console.log();
+    Logger.log();
+    Logger.info(`🛠  Building fresh story loader for ${packageName}`);
+    Logger.log();
 
     // 1. initialize expo project w/ name
     await runExpoCliAsync('init', [projectName, '-t', 'bare-minimum', '--no-install'], {
@@ -188,12 +192,12 @@ async function action({ platform, packageName }: Action) {
   fs.copyFileSync(path.resolve(templateRoot, 'App.js'), path.resolve(projectRoot, 'App.js'));
 
   // 4. yarn + install deps
-  console.log('🧶 Yarning...');
-  console.log();
+  Logger.log('🧶 Yarning...');
+  Logger.log();
   await spawnAsync('yarn', ['install'], { cwd: projectRoot });
 
-  console.log('✅ Done!');
-  console.log();
+  Logger.log('✅ Done!');
+  Logger.log();
 
   if (!platform) {
     const { selectedPlatform } = await inquirer.prompt({
@@ -210,13 +214,13 @@ async function action({ platform, packageName }: Action) {
     platform = selectedPlatform;
   }
 
-  console.log();
+  Logger.log();
 
   if (platform === 'web') {
   } else {
     const command = `run:${platform}`;
-    console.log(`🛠  Building for ${platform}...this may take a few minutes`);
-    console.log();
+    Logger.log(`🛠  Building for ${platform}...this may take a few minutes`);
+    Logger.log();
     runExpoCliAsync(command, [], { cwd: projectRoot });
   }
 
@@ -225,8 +229,7 @@ async function action({ platform, packageName }: Action) {
 
 export default (program: any) => {
   program
-    .command('run-stories')
+    .command('run-stories <packageName>')
     .option('-p, --platform <string>', 'Which platform we should run the app on')
-    .option('-n, --packageName <string>', 'The name of the package')
     .asyncAction(action);
 };
